Moved hashing concerns into the same project & namespace
