+++ src/SFA.DAS.EmployerApprenticeshipsService.Application.UnitTests/Queries/GetTransferTransactionDetailsTests/WhenIGetsTransferTransactionDetails.cs
@@ -9,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 
 namespace SFA.DAS.EAS.Application.UnitTests.Queries.GetTransferTransactionDetailsTests
