+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Commands/SendTransferConnectionInvitation/SendTransferConnectionInvitationCommandHandler.cs
@@ -2,6 +2,7 @@ using System.Threading.Tasks;
 using MediatR;
 using SFA.DAS.EAS.Domain.Data.Repositories;
 using SFA.DAS.EAS.Domain.Interfaces;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 
 namespace SFA.DAS.EAS.Application.Commands.SendTransferConnectionInvitation
