+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetEmployerAccountTransactions/GetEmployerAccountTransactionsHandler.cs
@@ -12,6 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 
 namespace SFA.DAS.EAS.Application.Queries.GetEmployerAccountTransactions
