+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/SendTransferConnectionInvitation/SendTransferConnectionInvitationQueryHandler.cs
@@ -8,6 +8,7 @@ using SFA.DAS.EAS.Application.Exceptions;
 using SFA.DAS.EAS.Domain.Models.TransferConnections;
 using SFA.DAS.EAS.Infrastructure.Data;
 using SFA.DAS.EAS.Infrastructure.Extensions;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 
 namespace SFA.DAS.EAS.Application.Queries.SendTransferConnectionInvitation
