+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/Data/AccountRepository.cs
@@ -12,6 +12,7 @@ using System.Collections.Generic;
 using System.Data;
 using System.Linq;
 using System.Threading.Tasks;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 using SFA.DAS.HashingService;
