+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/Hashing/IPublicHashingService.cs
@@ -1,6 +1,6 @@
 ﻿using SFA.DAS.HashingService;
 
-namespace SFA.DAS.EAS.Infrastructure.Interfaces
+namespace SFA.DAS.EAS.Infrastructure.Hashing
 {
     public interface IPublicHashingService : IHashingService
     {
