+++ src/SFA.DAS.EmployerApprenticeshipsService.Infrastructure/SFA.DAS.EAS.Infrastructure.csproj
@@ -311,8 +311,9 @@
     <Compile Include="Extensions\QueryableExtensions.cs" />
     <Compile Include="Extensions\QueryFutureEnumerableExtensions.cs" />
     <Compile Include="Features\IFeatureService.cs" />
-    <Compile Include="Interfaces\IPublicALEHashingService.cs" />
-    <Compile Include="Interfaces\IPublicHashingService.cs" />
+    <Compile Include="Hashing\PublicHashingService.cs" />
+    <Compile Include="Hashing\IPublicALEHashingService.cs" />
+    <Compile Include="Hashing\IPublicHashingService.cs" />
     <Compile Include="Mapping\Profiles\PaymentMappings.cs" />
     <Compile Include="Messaging\ServiceBusConnectionStringAttribute.cs" />
     <Compile Include="Messaging\ServiceBusConnectionStringPolicy.cs" />
