+++ src/SFA.DAS.EmployerApprenticeshipsService.TestCommon/DependencyResolution/DefaultRegistry.cs
@@ -9,8 +9,8 @@ using SFA.DAS.NLog.Logger;
 using SFA.DAS.Notifications.Api.Client;
 using StructureMap;
 using SFA.DAS.HashingService;
-using SFA.DAS.EAS.Application.Hashing;
 using SFA.DAS.EAS.Infrastructure.Authentication;
+using SFA.DAS.EAS.Infrastructure.Hashing;
 using SFA.DAS.EAS.Infrastructure.Interfaces;
 using SFA.DAS.Messaging.Interfaces;
